css/css-syntax/charset/page-utf16-css-bomless-utf16.html

<!doctype html>
<title>CSS charset: page utf-16, CSS bomless utf-16</title>
<link rel=help href="https://drafts.csswg.org/css-syntax-3/#determine-the-fallback-encoding">
<script src="/resources/testharness.js" charset=utf-8></script>
<script src="/resources/testharnessreport.js" charset=utf-8></script>
<link rel=stylesheet href="support/bomless-utf16.css">
<div id=log></div>
<div id=foo></div>
<script>
var elm = document.getElementById('foo');
var t = async_test();
onload = t.step_func(function(){
  assert_equals(getComputedStyle(elm, '').visibility, 'hidden');
  this.done();
});
</script>
